remove webpack from build
so we actually don't want to use webpack here. doing so bundles
dependencies together, and includes a production version of react which
greatly complicates requiring the project in a test environment.
instead, we can just use babel to transpile to ES5 and projects that
consume this module can deal with bundling it (they already do anyhow).

pr #4