use typescript to build the project
this switches over to using typescript. it

- adds and configures typescript to build the project
- converts files from .js -> .ts, .tsx
- adds @types/ typings for most of the libraries used
- adds type signatures to the code here
- configures eslint to work with typescript and adds typescript rules

overall this converts us fully over to using typescript to build this
project, and configures things so that type declarations are shipped
with the project when it is deployed to npm. so if we start using
typescript elsewhere (yes please!) we can use these types to fully
typecheck our use of the useCourseSearch hook.

pr #10